requirements.txt
openai-whisper
torch
numpy
ffmpeg-python
pysrt
deep-translator
requests